+++ refena_flutter/lib/refena_flutter.dart
@@ -1,10 +1,10 @@
 export 'package:refena/refena.dart';
 export 'package:refena_flutter/src/consumer.dart';
 export 'package:refena_flutter/src/extension.dart';
+export 'package:refena_flutter/src/family_view_model_builder.dart';
 export 'package:refena_flutter/src/graph/graph_page.dart';
 export 'package:refena_flutter/src/mixin.dart';
 export 'package:refena_flutter/src/scope.dart' show RefenaScope;
 export 'package:refena_flutter/src/tracing/tracing_page.dart';
 export 'package:refena_flutter/src/view_model_builder.dart';
-export 'package:refena_flutter/src/view_model_param_builder.dart';
 export 'package:refena_flutter/src/widget_rebuildable.dart';